% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgi.R
\name{load_wgi_indicator}
\alias{load_wgi_indicator}
\title{load_wgi_indicator}
\usage{
load_wgi_indicator(indicatorname)
}
\arguments{
\item{indicatorname}{a indicator need to extract : VAB, PSV,GEN, RGQ, ROL, COF.}
}
\value{
data.frame
}
\description{
load_wgi_indicator
}
\examples{
load_wgi_indicator("VAB")
}
