% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgi.R
\name{read_wgi}
\alias{read_wgi}
\title{read_wgi}
\usage{
read_wgi(
  startyear = NULL,
  endyear = NULL,
  country = NULL,
  indicator = NULL,
  variable = NULL,
  rm.na = TRUE
)
}
\arguments{
\item{startyear}{start year}

\item{endyear}{end year}

\item{country}{vector of the countries , default is all the countries present in WGI.}

\item{indicator}{is vector of string values: VAB= Voice and Accountability,
PSV="Political Stability No Violence, GEN=Government Effectiveness", RGQ=Regulatory Quality, ROL=Rule of Law, and COF=Control of Corruption.}

\item{variable}{is string value such as : "Estimate", "StdErr" ,  "NumSrc" ,  "Rank" ,    "Lower", and    "Upper"}

\item{rm.na}{if it is FALSE preserve missing values, default is TRUE}
}
\value{
data.frame
}
\description{
read_wgi
}
\examples{
wgi_df<-read_wgi(startyear=2021, endyear=2022)
}
